Add files via upload
The Company is a fictional company that has administered a survey to all of its employees. They have recently implemented some efforts to improve DEI within the company, and this survey is for management to gauge the employees feelings about these efforts. 

As a data analyst, I’ve been hired to answer the following questions management has given me:

Do the employees who belong to historically marginalized demographics (for example: women, people in the LGBTQ community, people of color, etc.) feel company efforts in DEI actually work?
Are there any groups that feel a negative impact from the company’s efforts. If so, who?